Написана программа для сортировки записей в файлы по заданным IP адресам.
